+++ futures-core/src/task.rs
@@ -1,0 +1,11 @@
+//! Task notification.
+
+pub use core::task::{
+    Context, Poll, Executor,
+    Waker, LocalWaker, UnsafeWake,
+    SpawnErrorKind, SpawnObjError, SpawnLocalObjError,
+};
+
+if_std! {
+    pub use std::task::{Wake, local_waker, local_waker_from_nonlocal};
+}
