+++ futures-util/src/task/mod.rs
@@ -1,0 +1,9 @@
+//! Task notification
+
+mod context;
+pub use self::context::ContextExt;
+
+#[cfg_attr(feature = "nightly", cfg(target_has_atomic = "ptr"))]
+mod atomic_waker;
+#[cfg_attr(feature = "nightly", cfg(target_has_atomic = "ptr"))]
+pub use self::atomic_waker::AtomicWaker;
